requirements.txt
# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# LINE Bot SDK
line-bot-sdk==3.11.0

# Database
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis / Cache
redis==5.0.4
celery==5.4.0
flower==2.0.1  # Celery 監控

# Image Processing
Pillow==10.3.0
httpx==0.27.0  # 非同步 HTTP 請求

# Environment / Config
python-dotenv==1.0.1
pydantic-settings==2.2.1

# Security / Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
aiofiles==23.2.1
boto3==1.34.95   # AWS S3 / Rekognition
loguru==0.7.2

# Payment (藍新金流)
requests==2.31.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
